+++ crates/ika-core/src/dwallet_mpc/submit_to_consensus.rs
@@ -1,7 +1,8 @@
 use crate::consensus_adapter::SubmitToConsensus;
-use crate::dwallet_mpc::dwallet_mpc_service::EpochStoreSubmitToConsensus;
 use ika_types::error::IkaResult;
 use ika_types::messages_consensus::ConsensusTransaction;
+use std::sync::Arc;
+use crate::authority::authority_per_epoch_store::AuthorityPerEpochStore;
 
 #[async_trait::async_trait]
 pub trait DWalletMPCSubmitToConsensus: Sync + Send + 'static {
@@ -14,5 +15,22 @@ impl DWalletMPCSubmitToConsensus for EpochStoreSubmitToConsensus {
         self.consensus_adapter
             .submit_to_consensus(transactions, &self.epoch_store)
             .await
+    }
+}
+
+pub struct EpochStoreSubmitToConsensus {
+    pub(crate) epoch_store: Arc<AuthorityPerEpochStore>,
+    pub(crate) consensus_adapter: Arc<dyn SubmitToConsensus>,
+}
+
+impl EpochStoreSubmitToConsensus {
+    pub fn new(
+        epoch_store: Arc<AuthorityPerEpochStore>,
+        consensus_adapter: Arc<dyn SubmitToConsensus>,
+    ) -> Self {
+        Self {
+            epoch_store,
+            consensus_adapter,
+        }
     }
 }
